Fix day 8 test

+++ test/day8.cpp
@@ -5,7 +5,7 @@
 
 TEST(Day8Test, Part1Example) {
   InputFile in("day8_ex");
-  int result = day8_part1(in.lines, in.nlines);
+  long result = day8_part1(in.lines, in.nlines);
   EXPECT_EQ(result, 2);
   InputFile in2("day8_ex2");
   result = day8_part1(in2.lines, in2.nlines);
@@ -14,18 +14,18 @@ TEST(Day8Test, Part1Example) {
 
 TEST(Day8Test, Part1Input) {
   InputFile in("day8");
-  int result = day8_part1(in.lines, in.nlines);
+  long result = day8_part1(in.lines, in.nlines);
   EXPECT_EQ(result, get_solution(8, 1));
 }
 
 TEST(Day8Test, Part2Example) {
   InputFile in("day8_ex3");
-  int result = day8_part2(in.lines, in.nlines);
+  long result = day8_part2(in.lines, in.nlines);
   EXPECT_EQ(result, 6);
 }
 
 TEST(Day8Test, Part2Input) {
   InputFile in("day8");
-  int result = day8_part2(in.lines, in.nlines);
+  long result = day8_part2(in.lines, in.nlines);
   EXPECT_EQ(result, get_solution(8, 2));
 }
